FEATURE: Filtering of connection attempts using DNSBL.
Adds a new configuation option in the service config `resolverAddress`.

This address is used as part of the connection handler in the SSH strategy.

When a new connection is received, the source address is checked against
the Spamhaus Zen DNSBL, and if it is presented in the DROP list, the
connection is closed. The "resolverAddress" defines the recursive DNS
server used to make the DNS lookup. Per Spamhaus terms of service, this
cannot be a public resolver (such as Google, CloudFlare, or Quad9), and
will return an error (but still process the connection) if used in this way.

Future enhancements may include local caching of known bad addresses.